+++ src/plugin/__tests__/mount.spec.ts
@@ -1,7 +1,7 @@
 import { buildElementPlugin } from "../element";
 import { createElementSpec } from "../elementSpec";
+import { createNoopElement } from "../helpers/test";
 import type { CustomField } from "../types/Element";
-import { createNoopElement } from "./helpers";
 
 describe("mount", () => {
   describe("fieldView typesafety", () => {
