+++ src/plugin/element.ts
@@ -2,7 +2,7 @@ import OrderedMap from "orderedmap";
 import type { NodeSpec, Schema } from "prosemirror-model";
 import type { EditorState, Transaction } from "prosemirror-state";
 import type { FieldNameToValueMap } from "./fieldViews/helpers";
-import { buildCommands, defaultPredicate } from "./helpers";
+import { buildCommands, defaultPredicate } from "./helpers/prosemirror";
 import { createNodesForFieldValues } from "./nodeSpec";
 import { createPlugin } from "./plugin";
 import type { ElementSpec, FieldSpec } from "./types/Element";
