+++ src/plugin/helpers/prosemirror.ts
@@ -1,9 +1,9 @@
-import type { Node } from "prosemirror-model";
+import type { Node, Schema } from "prosemirror-model";
+import { DOMParser, DOMSerializer } from "prosemirror-model";
 import type { EditorState, Transaction } from "prosemirror-state";
 import { AllSelection } from "prosemirror-state";
 import type { EditorView } from "prosemirror-view";
 import { Decoration, DecorationSet } from "prosemirror-view";
-import type { SetMedia } from "../elements/image/imageElement";
 
 type NodesBetweenArgs = [Node, number, Node, number];
 export type Commands = ReturnType<typeof buildCommands>;
@@ -205,85 +205,30 @@ const createDecorations = (name: string) => (state: EditorState) => {
   return DecorationSet.create(state.doc, decorations);
 };
 
-const onGridMessage = (setMedia: SetMedia, modal: HTMLElement) => ({
-  data,
-}: {
-  data: {
-    image: {
-      data: {
-        id: string;
-      };
-    };
-    crop: {
-      data: {
-        specification: {
-          uri: string;
-        };
-        assets: Array<{ secureUrl: string }>;
-      };
-    };
-  };
-}) => {
-  modal.style.display = "None";
-  setMedia(
-    data.image.data.id,
-    data.crop.data.specification.uri,
-    data.crop.data.assets.map((_) => _.secureUrl)
-  );
+const createParsers = <S extends Schema>(schema: S) => {
+  const parser = DOMParser.fromSchema(schema);
+  const serializer = DOMSerializer.fromSchema(schema);
+  return { parser, serializer };
 };
 
-const onSelectImage = (setMedia: SetMedia) => {
-  const modal = document.querySelector(".modal") as HTMLElement;
-  modal.style.display = "Inherit";
-
-  (document.querySelector(
-    ".modal__body iframe"
-  ) as HTMLIFrameElement).src = `https://media.test.dev-gutools.co.uk/`;
-
-  const listener = onGridMessage(setMedia, modal);
-
-  window.addEventListener("message", listener, {
-    once: true,
-  });
-
-  document.querySelector(".modal__dismiss")?.addEventListener(
-    "click",
-    () => {
-      window.removeEventListener("message", listener);
-      modal.style.display = "None";
-    },
-    { once: false }
-  );
+const docToHtml = (serializer: DOMSerializer, doc: Node) => {
+  const dom = serializer.serializeFragment(doc.content);
+  const e = document.createElement("div");
+  e.appendChild(dom);
+  return e.innerHTML;
 };
 
-const onCropImage = (mediaId: string, setMedia: SetMedia) => {
-  const modal = document.querySelector(".modal") as HTMLElement;
-  modal.style.display = "Inherit";
-
-  (document.querySelector(
-    ".modal__body iframe"
-  ) as HTMLIFrameElement).src = `https://media.test.dev-gutools.co.uk/images/${mediaId}`;
-
-  const listener = onGridMessage(setMedia, modal);
-
-  window.addEventListener("message", listener, {
-    once: true,
-  });
-
-  document.querySelector(".modal__dismiss")?.addEventListener(
-    "click",
-    () => {
-      window.removeEventListener("message", listener);
-      modal.style.display = "None";
-    },
-    { once: false }
-  );
+const htmlToDoc = (parser: DOMParser, html: string) => {
+  const dom = document.createElement("div");
+  dom.innerHTML = html;
+  return parser.parse(dom);
 };
 
 export {
   buildCommands,
   defaultPredicate,
   createDecorations,
-  onSelectImage,
-  onCropImage,
+  createParsers,
+  docToHtml,
+  htmlToDoc,
 };
